REFACTOR - junta o ListarQuartosController no QuartosController
